fix(session): preserve liveness safeguards in reconcile lock path (#654)
Addresses PR #655 round-2 findings:
- codex P1 x2: restore session context + staleness check for daemon.pid / lock-file PID
- gemini-code-assist high: parse lock file via serde_json, not substring search

Choice A (restore helpers) was taken. Sub-agent logic in session_cmds_reconcile.rs
was replaced with ToolLiveness::has_live_process from csa-process.
ToolLiveness was strengthened to:
1. Use serde_json for robust PID extraction.
2. Require BOTH context match AND staleness for lock files (&& instead of ||).
3. Exclude reconcile.lock from tool discovery.

Tests added for PID reuse (via spawned sleep), stale locks, and complex JSON.
Tests moved to separate session_cmds_reconcile_tests.rs to satisfy 800-line limit.
Fixed existing tests in csa-process and cli-sub-agent broken by stronger context check.
