+++ crates/cli-sub-agent/src/session_cmds_reconcile.rs
@@ -6,11 +6,11 @@ use std::os::fd::AsRawFd;
 use std::path::{Path, PathBuf};
 use tracing::{info, warn};
 
+use csa_process::ToolLiveness;
 use csa_session::{
     MetaSessionState, SessionPhase, SessionResult, get_session_dir, load_result, load_session,
 };
 
-const RECONCILE_LOCK_NAME: &str = "reconcile";
 type PersistSessionFn<'a> = dyn Fn(&Path, &MetaSessionState) -> Result<()> + 'a;
 
 #[rustfmt::skip]
@@ -103,7 +103,7 @@ where
     if !matches!(session.phase, SessionPhase::Active) {
         return Ok(DeadActiveSessionReconciliation::NoChange);
     }
-    if session_has_live_tool_process(session_dir) {
+    if ToolLiveness::has_live_process(session_dir) {
         return Ok(DeadActiveSessionReconciliation::NoChange);
     }
     let result_path = session_dir.join(csa_session::result::RESULT_FILE_NAME);
@@ -261,7 +261,7 @@ fn retire_if_dead_with_result_impl(
     if !matches!(session.phase, SessionPhase::Active) {
         return Ok(false);
     }
-    if session_has_live_tool_process(session_dir)
+    if ToolLiveness::has_live_process(session_dir)
         || load_result(project_root, session_id)?.is_none()
     {
         return Ok(false);
@@ -337,68 +337,6 @@ fn acquire_reconcile_lock(
     {
         let _ = trigger;
         Ok(Some((session_dir, ReconcileLock { file })))
-    }
-}
-
-fn session_has_live_tool_process(session_dir: &Path) -> bool {
-    if read_session_daemon_pid(session_dir).is_some_and(is_process_alive) {
-        return true;
-    }
-
-    let locks_dir = session_dir.join("locks");
-    let Ok(entries) = fs::read_dir(&locks_dir) else {
-        return false;
-    };
-
-    entries.flatten().any(|entry| {
-        let path = entry.path();
-        if path.extension().is_none_or(|ext| ext != "lock") {
-            return false;
-        }
-        if path
-            .file_stem()
-            .and_then(|stem| stem.to_str())
-            .is_some_and(|stem| stem == RECONCILE_LOCK_NAME)
-        {
-            return false;
-        }
-        let Ok(content) = fs::read_to_string(&path) else {
-            return false;
-        };
-        extract_lock_pid(&content).is_some_and(is_process_alive)
-    })
-}
-
-#[rustfmt::skip]
-fn read_session_daemon_pid(session_dir: &Path) -> Option<u32> { fs::read_to_string(session_dir.join("daemon.pid")).ok()?.trim().parse::<u32>().ok() }
-
-fn extract_lock_pid(lock_content: &str) -> Option<u32> {
-    let pid_key_pos = lock_content.find("\"pid\"")?;
-    let tail = &lock_content[pid_key_pos..];
-    let colon_pos = tail.find(':')?;
-    let number = tail[colon_pos + 1..]
-        .chars()
-        .skip_while(|ch| ch.is_ascii_whitespace())
-        .take_while(|ch| ch.is_ascii_digit())
-        .collect::<String>();
-    number.parse::<u32>().ok()
-}
-
-fn is_process_alive(pid: u32) -> bool {
-    #[cfg(unix)]
-    {
-        // SAFETY: `kill(pid, 0)` performs an existence probe without sending a signal.
-        let ret = unsafe { libc::kill(pid as libc::pid_t, 0) };
-        if ret == 0 {
-            return true;
-        }
-        let errno = std::io::Error::last_os_error().raw_os_error();
-        errno == Some(libc::EPERM)
-    }
-
-    #[cfg(not(unix))]
-    {
-        std::path::Path::new(&format!("/proc/{pid}/stat")).exists()
     }
 }
 
@@ -503,297 +441,5 @@ fn format_optional_file_mtime(path: &Path) -> Option<String> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::session_cmds_daemon::{WaitReconciliationOutcome, handle_session_wait_with_hooks};
-    use crate::test_env_lock::TEST_ENV_LOCK;
-    use chrono::Utc;
-    use csa_session::{
-        SessionPhase, SessionResult, create_session, get_session_dir, load_result, load_session,
-        save_result,
-    };
-    use tempfile::tempdir;
-
-    struct EnvVarGuard {
-        key: &'static str,
-        original: Option<String>,
-    }
-
-    impl EnvVarGuard {
-        fn set(key: &'static str, value: impl AsRef<std::ffi::OsStr>) -> Self {
-            let original = std::env::var(key).ok();
-            // SAFETY: test-scoped env mutation guarded by a process-wide mutex.
-            unsafe { std::env::set_var(key, value) };
-            Self { key, original }
-        }
-    }
-
-    impl Drop for EnvVarGuard {
-        fn drop(&mut self) {
-            // SAFETY: test-scoped env mutation guarded by a process-wide mutex.
-            unsafe {
-                match self.original.as_deref() {
-                    Some(value) => std::env::set_var(self.key, value),
-                    None => std::env::remove_var(self.key),
-                }
-            }
-        }
-    }
-
-    struct SessionTestEnv {
-        _env_lock: std::sync::MutexGuard<'static, ()>,
-        _home_guard: EnvVarGuard,
-        _state_guard: EnvVarGuard,
-    }
-
-    impl SessionTestEnv {
-        fn new(td: &tempfile::TempDir) -> Self {
-            let env_lock = TEST_ENV_LOCK.lock().expect("session env lock poisoned");
-            let state_home = td.path().join("xdg-state");
-            std::fs::create_dir_all(&state_home).expect("create state home");
-            let home_guard = EnvVarGuard::set("HOME", td.path());
-            let state_guard = EnvVarGuard::set("XDG_STATE_HOME", &state_home);
-            Self {
-                _env_lock: env_lock,
-                _home_guard: home_guard,
-                _state_guard: state_guard,
-            }
-        }
-    }
-
-    fn make_result(status: &str, exit_code: i32) -> SessionResult {
-        let now = Utc::now();
-        SessionResult {
-            status: status.to_string(),
-            exit_code,
-            summary: "summary".to_string(),
-            tool: "codex".to_string(),
-            started_at: now,
-            completed_at: now,
-            events_count: 0,
-            artifacts: Vec::new(),
-            peak_memory_mb: None,
-        }
-    }
-
-    #[test]
-    fn ensure_terminal_result_for_dead_active_session_leaves_state_unchanged_on_transition_failure()
-    {
-        let td = tempdir().expect("tempdir");
-        let _env = SessionTestEnv::new(&td);
-        let project = td.path();
-
-        let session = create_session(project, Some("transition-failure"), None, None).unwrap();
-        let session_id = session.meta_session_id;
-        let session_dir = get_session_dir(project, &session_id).unwrap();
-        let state_path = session_dir.join("state.toml");
-        let cooldown_path = session_dir.parent().unwrap().join("cooldown-marker.toml");
-        let original_state = fs::read_to_string(&state_path).unwrap();
-
-        let reconciled = ensure_terminal_result_for_dead_active_session_impl(
-            project,
-            &session_id,
-            "session list",
-            &session_dir,
-            |_| {},
-            |session| {
-                session.phase = SessionPhase::Retired;
-            },
-            &persist_session_state_atomically,
-        )
-        .unwrap();
-
-        assert_eq!(reconciled, DeadActiveSessionReconciliation::NoChange);
-        let persisted = load_session(project, &session_id).unwrap();
-        assert_eq!(persisted.phase, SessionPhase::Active);
-        assert_eq!(persisted.termination_reason, None);
-        assert_eq!(fs::read_to_string(&state_path).unwrap(), original_state);
-        assert!(
-            load_result(project, &session_id).unwrap().is_none(),
-            "synthetic result should be removed after a transition failure"
-        );
-        assert!(
-            !cooldown_path.exists(),
-            "cooldown marker should not be written for a rolled-back reconciliation"
-        );
-    }
-
-    #[test]
-    fn ensure_terminal_result_for_dead_active_session_removes_synthetic_result_on_save_failure() {
-        let td = tempdir().expect("tempdir");
-        let _env = SessionTestEnv::new(&td);
-        let project = td.path();
-
-        let session = create_session(project, Some("save-failure"), None, None).unwrap();
-        let session_id = session.meta_session_id;
-        let session_dir = get_session_dir(project, &session_id).unwrap();
-        let state_path = session_dir.join("state.toml");
-        let cooldown_path = session_dir.parent().unwrap().join("cooldown-marker.toml");
-        let original_state = fs::read_to_string(&state_path).unwrap();
-        let persist_fail = |_: &Path, _: &MetaSessionState| -> Result<()> { Err(anyhow!("boom")) };
-
-        let err = ensure_terminal_result_for_dead_active_session_impl(
-            project,
-            &session_id,
-            "session list",
-            &session_dir,
-            |_| {},
-            |_| {},
-            &persist_fail,
-        )
-        .unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("Failed to persist retired orphaned session state"),
-            "unexpected error: {err:#}"
-        );
-        let persisted = load_session(project, &session_id).unwrap();
-        assert_eq!(persisted.phase, SessionPhase::Active);
-        assert_eq!(persisted.termination_reason, None);
-        assert_eq!(fs::read_to_string(&state_path).unwrap(), original_state);
-        assert!(
-            load_result(project, &session_id).unwrap().is_none(),
-            "synthetic result should be removed after a state persistence failure"
-        );
-        assert!(
-            !cooldown_path.exists(),
-            "cooldown marker should not be written when reconciliation state persistence fails"
-        );
-    }
-
-    #[test]
-    fn retire_if_dead_with_result_leaves_state_unchanged_on_save_failure() {
-        let td = tempdir().expect("tempdir");
-        let _env = SessionTestEnv::new(&td);
-        let project = td.path();
-
-        let session = create_session(project, Some("retire-save-failure"), None, None).unwrap();
-        let session_id = session.meta_session_id;
-        save_result(project, &session_id, &make_result("success", 0)).unwrap();
-        let session_dir = get_session_dir(project, &session_id).unwrap();
-        let state_path = session_dir.join("state.toml");
-        let original_state = fs::read_to_string(&state_path).unwrap();
-        let persist_fail = |_: &Path, _: &MetaSessionState| -> Result<()> { Err(anyhow!("boom")) };
-
-        let err = retire_if_dead_with_result_impl(
-            project,
-            &session_id,
-            "session list",
-            &session_dir,
-            &persist_fail,
-        )
-        .unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("Failed to persist retired session state"),
-            "unexpected error: {err:#}"
-        );
-        let persisted = load_session(project, &session_id).unwrap();
-        assert_eq!(persisted.phase, SessionPhase::Active);
-        assert_eq!(persisted.termination_reason, None);
-        assert_eq!(fs::read_to_string(&state_path).unwrap(), original_state);
-        let result = load_result(project, &session_id).unwrap().unwrap();
-        assert_eq!(result.status, "success");
-        assert_eq!(result.exit_code, 0);
-    }
-
-    #[test]
-    fn ensure_terminal_result_for_dead_active_session_is_noop_when_reconcile_lock_is_held() {
-        let td = tempdir().expect("tempdir");
-        let _env = SessionTestEnv::new(&td);
-        let project = td.path();
-
-        let session = create_session(project, Some("lock-held"), None, None).unwrap();
-        let session_id = session.meta_session_id;
-        let (_session_dir, _lock) = acquire_reconcile_lock(project, &session_id, "unit-test")
-            .unwrap()
-            .expect("lock should be acquired for setup");
-
-        let reconciled =
-            ensure_terminal_result_for_dead_active_session(project, &session_id, "session list")
-                .unwrap();
-
-        assert_eq!(reconciled, DeadActiveSessionReconciliation::NoChange);
-        assert!(load_result(project, &session_id).unwrap().is_none());
-        let persisted = load_session(project, &session_id).unwrap();
-        assert_eq!(persisted.phase, SessionPhase::Active);
-        assert_eq!(persisted.termination_reason, None);
-    }
-
-    #[test]
-    fn persist_new_result_file_removes_partial_file_when_write_fails() {
-        let td = tempdir().expect("tempdir");
-        let result_path = td.path().join("result.toml");
-
-        let err = persist_new_result_file_with_writer(
-            &result_path,
-            "status = \"failure\"\n",
-            |_| {},
-            |file, _contents| {
-                file.write_all(b"partial")?;
-                Err(std::io::Error::other("boom"))
-            },
-        )
-        .unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("Failed to write or sync synthetic result"),
-            "unexpected error: {err:#}"
-        );
-        assert!(
-            !result_path.exists(),
-            "partial synthetic result should be removed after write failure"
-        );
-    }
-
-    #[test]
-    fn handle_session_wait_marks_late_real_result_completion_as_non_synthetic() {
-        let td = tempdir().expect("tempdir");
-        let _env = SessionTestEnv::new(&td);
-        let project = td.path();
-
-        let session =
-            create_session(project, Some("wait-late-real-result"), None, Some("codex")).unwrap();
-        let session_id = session.meta_session_id;
-        let late_result = SessionResult {
-            summary: "real terminal result".to_string(),
-            ..make_result("success", 0)
-        };
-        let mut emitted_completion: Option<(String, String, i32, bool)> = None;
-
-        let exit_code = handle_session_wait_with_hooks(
-            session_id.clone(),
-            Some(project.to_string_lossy().into_owned()),
-            5,
-            |project_root, current_session_id, trigger| {
-                let reconciled = ensure_terminal_result_for_dead_active_session_with_before_write(
-                    project_root,
-                    current_session_id,
-                    trigger,
-                    |_| {
-                        save_result(project_root, current_session_id, &late_result)
-                            .expect("persist late real result");
-                    },
-                )?;
-                Ok(WaitReconciliationOutcome {
-                    result_became_available: reconciled.result_became_available(),
-                    synthetic: reconciled.synthesized_failure(),
-                })
-            },
-            |sid, status, exit_code, synthetic, _mirror_to_stdout| {
-                emitted_completion =
-                    Some((sid.to_string(), status.to_string(), exit_code, synthetic));
-            },
-        )
-        .unwrap();
-
-        assert_eq!(exit_code, 0);
-        assert_eq!(
-            emitted_completion,
-            Some((session_id, "success".to_string(), 0, false))
-        );
-    }
-}
+#[path = "session_cmds_reconcile_tests.rs"]
+mod tests;
